+++ CHANGELOG.rst
@@ -10,15 +10,9 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
-- Added ``dtool config ecs ls`` command to list ECS base URIs that have been
-  configured
-
 
 Changed
 ^^^^^^^
-
-- The ``dtool config azure ls`` command now returns base URIs rather than
-  container names
 
 
 Deprecated
@@ -35,6 +29,24 @@ Fixed
 
 Security
 ^^^^^^^^
+
+
+[0.4.0] - 2019-07-12
+--------------------
+
+Added
+^^^^^
+
+- Added ``dtool config ecs ls`` command to list ECS base URIs that have been
+  configured
+
+
+Changed
+^^^^^^^
+
+- The ``dtool config azure ls`` command now returns base URIs rather than
+  container names
+
 
 [0.3.0] - 2019-05-14
 --------------------